fixed error when a.dat is not exist.
git-svn-id: https://tdiary.svn.sourceforge.net/svnroot/tdiary/trunk/core@3371 7f22e88f-374d-0410-998f-c91420d97ba2

+++ plugin/90migrate.rb
@@ -87,8 +87,10 @@ unless @conf.tdiary_version
 
 	# other files
 	a_dat = @conf['a.path'] || "#{dir}/a.dat"
-	t = File::open( a_dat ){|f| f.read}
-	File::open( a_dat, 'wb' ){|f| f.write( @conf.migrate_to_utf8( t ) )}
+	if File.exist?( a_dat ) then
+		t = File::open( a_dat ){|f| f.read}
+		File::open( a_dat, 'wb' ){|f| f.write( @conf.migrate_to_utf8( t ) )}
+	end
 
 	# remove ruby/erb cache files
 	Dir["#{dir}/*.rb"].each{|f| FileUtils.rm_f( f )}
